requirements.txt
deepface==0.0.80
opencv-python==4.8.1.78
paddleocr==2.6.0.1
pandas==2.1.4
Pillow==10.1.0
PyYAML==6.0.1
SQLAlchemy==2.0.23
streamlit==1.29.0
ultralytics==8.0.227

python-dotenv~=1.0.0
fuzzywuzzy~=0.18.0
requests~=2.31.0